% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{criterion_likelihood}
\alias{criterion_likelihood}
\title{Calculate Model Likelihood indicators (NLL, AIC, BIC)}
\usage{
criterion_likelihood(freq_mat, prob_mat, std_params)
}
\description{
Calculate Model Likelihood indicators (NLL, AIC, BIC)
}
