% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{data_info}
\alias{data_info}
\title{Intelligently scan the dataset and extract subject-level information.}
\usage{
data_info(df, colnames = NULL)
}
\description{
Intelligently scan the dataset and extract subject-level information.
}
