% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{default_shell_run_m_option}
\alias{default_shell_run_m_option}
\title{Default options for shell_run_m}
\usage{
default_shell_run_m_option()
}
\value{
A list of shell-level simulation and display options.
}
\description{
Default options for shell_run_m
}
