% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{fit_meta_d_logL}
\alias{fit_meta_d_logL}
\title{Likelihood function for fitting meta-d}
\usage{
fit_meta_d_logL(x, parameters)
}
\description{
Likelihood function for fitting meta-d
}
\author{
Maniscalco and Lau. Ported to R by Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
