% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{fit_meta_d_MLE}
\alias{fit_meta_d_MLE}
\title{Fit Type 2 SDT using Maximum Likelihood Estimation.}
\usage{
fit_meta_d_MLE(nR_S1, nR_S2, s = 1, add_constant = TRUE)
}
\arguments{
\item{nR_S1}{Responses to S1 stimulus. See below for advice.}

\item{nR_S2}{Responses to S2 stimulus. See below for advice.}

\item{s}{Standard deviation. Defaults to 1.}

\item{add_constant}{Adds a small constant to the data (1/number of possible responses) to account for 0 or 1 values. Defaults to TRUE for ease of use across multiple datasets.}
}
\description{
This uses data in count format.
}
\references{
Maniscalco, B., & Lau, H. (2012). A signal detection theoretic approach for estimating metacognitive sensitivity from confidence ratings. Consciousness and Cognition. http://dx.doi.org/10.1016/j.concog.2011.09.021
}
\author{
Maniscalco & Lau. Ported to R by Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
