% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{fit_meta_d_sse}
\alias{fit_meta_d_sse}
\title{Function for calculating meta-d' using SSE}
\usage{
fit_meta_d_sse(nR_S1, nR_S2, s = 1, d_min = -5, d_max = 5,
  d_grain = 0.01, add_constant = TRUE)
}
\description{
Function for calculating meta-d' using SSE
}
