% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{fit_metad_minus}
\alias{fit_metad_minus}
\title{Internal function for fitting meta_d_minus}
\usage{
fit_metad_minus(x0, th, hm, fm)
}
\arguments{
\item{x0}{Starting parameters.}

\item{th}{Theta}

\item{hm}{Hit rate for negative responses}

\item{fm}{False positive rate for negative responses}
}
\description{
Internal function for fitting meta_d_minus
}
