% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{fit_metad_plus}
\alias{fit_metad_plus}
\title{Internal function for fitting meta_d_plus}
\usage{
fit_metad_plus(x0, th, hp, fp)
}
\arguments{
\item{x0}{Starting parameters.}

\item{th}{Theta.}

\item{hp}{Hit rate for positive responses}

\item{fp}{False positive rate for positive responses.}
}
\description{
Internal function for fitting meta_d_plus
}
