% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\\name{info_data}
\\alias{info_data}
\title{Intelligently scan the dataset and extract subject-level information.}
\usage{
info_data(df, colnames = NULL)
}
\description{
Intelligently scan the dataset and extract subject-level information.
}

