% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matrix_freq}
\alias{matrix_freq}
\title{Calculate the frequency matrix for Signal Detection Theory (C++ core)}
\usage{
matrix_freq(stim, resp, conf = NULL)
}
\arguments{
\item{stim}{The numeric vector indicating the signal type.}

\item{resp}{The numeric vector indicating the decision or response.}

\item{conf}{The numeric vector indicating the confidence level (optional).}
}
\description{
Calculate the frequency matrix for Signal Detection Theory (C++ core)
}
