% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_run_m.R
\name{plot.metaSDT.shell_run_m}
\alias{plot.metaSDT.shell_run_m}
\title{Plot a shell_run_m simulation}
\usage{
\method{plot}{metaSDT.shell_run_m}(x, show = TRUE, ...)
}
\arguments{
\item{x}{A \code{"metaSDT.shell_run_m"} object.}

\item{show}{Whether to print the plot.}

\item{...}{Unused.}
}
\value{
A ggplot object, invisibly.
}
\description{
Plot a shell_run_m simulation
}
