% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shell_run_m.R
\name{print.metaSDT.shell_run_m}
\alias{print.metaSDT.shell_run_m}
\title{Print a shell_run_m simulation}
\usage{
\method{print}{metaSDT.shell_run_m}(x, ...)
}
\arguments{
\item{x}{A \code{"metaSDT.shell_run_m"} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a shell_run_m simulation
}
