% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{rating_sdt}
\alias{rating_sdt}
\title{Type 1 SDT for rating data}
\usage{
rating_sdt(nr_s1, nr_s2, add_constant = FALSE)
}
\arguments{
\item{add_constant}{Adds a small constant to every cell to account for
boundaries - i.e. log-linear correction. Default = TRUE.}

\item{nR_S1}{Responses to S1 stimulus. See below for advice.}

\item{nR_S2}{Responses to S2 stimulus. See below for advice.}
}
\description{
Provides a Type-1 SDT analysis of data from a rating experiment in which
observers discriminate between two response alternatives and provide ratings
of confidence in their judgements.
}
\details{
The expected input is two vectors, one for responses to each stimulus,
encoding the observers response and confidence. For example, for two stimului
labelled A and B, with three confidence ratings, participants could respond
to stimulus A as follows:

Response: A, rating: 3, count: 60

Response: A, rating: 2, count: 30

Response: A, rating: 1, count: 10

Response: B, rating: 1, count: 7

Response: B, rating: 2, count: 4

Response: B, rating: 3, count: 1

The appropriate vector would be nR_S1 <- c(60,30,10,7,4,1)

For stimulus B, we would have the respective vector for responses to stimulus
B, eg:

Response: A, rating: 3, count: 4

Response: A, rating: 2, count: 6

Response: A, rating: 1, count: 11

Response: B, rating: 1, count: 13

Response: B, rating: 2, count: 23

Response: B, rating: 3, count: 61

nR_S2 <- c(4,6,11,13,23,61)

The helper function \code{sdt_counts} can be used to get the data into the
right format.

The output is a data frame with various Type-1 measures, one for each
criterion.
}
\author{
Matt Craddock \email{matt@mattcraddock.com}
}
