% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{sdt_counts}
\alias{sdt_counts}
\title{Convert to counts}
\usage{
sdt_counts(df, stimulus = NULL, response = NULL, ...)
}
\description{
Convert to counts
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
