% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdt_functions.R
\name{type_1_sdt}
\alias{type_1_sdt}
\title{Type 1 SDT for a 2AFC design}
\usage{
type_1_sdt(df, stimulus = NULL, response = NULL, counts = total,
  confidence = NULL, s = 1, add_constant = TRUE)
}
\arguments{
\item{df}{Raw data frame in long format. Assumes counts already calculated. Expects the response column to be 1 or 0 for reported yes or reported no.}

\item{stimulus}{Column name for levels of the stimulus. Should be bare, unquoted. e.g. (stimulus = stimulus)}

\item{response}{Col}
}
\description{
This calculates standard, type 1 SDT measures.
}
\author{
Matt Craddock, \email{m.p.craddock@leeds.ac.uk}
}
